backend-backup/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.13.0
stripe==7.8.0
telnyx==2.1.1
requests==2.31.0
pydantic==2.5.0
email-validator==2.1.0
